+++ packages/client/src/app/App.tsx
@@ -5,8 +5,8 @@ import { FORUM_ROUTE, MAIN_ROUTE, SIGN_IN_ROUTE, SIGN_OUT_ROUTE } from '../utils
 import { Main } from '../pages/main'
 import { SignIn } from '../pages/sign-in'
 import { SignUp } from '../pages/sign-up'
-import { ForumPick } from '../pages/forum/forum-pick'
-import { ForumChat } from '../pages/forum/forum-chat'
+import { ForumPick } from '../pages/forum/forumPick'
+import { ForumChat } from '../pages/forum/forumChat'
 
 function RequireAuth({ children }: { children: JSX.Element }) {
   // Когда будем подключать аунтификацию - будет использоваться обертка, для проверки авторизован ли пользователь,
